Chatbot/requirements.txt
flask==3.0.2
openai>=1.0.0
pandas==2.2.0
python-dotenv==1.0.1
streamlit==1.32.0 